Added `MemoryCacheHandler`, a cache handler that simply stores the token info in memory as an instance attribute of this class.
