+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-// Add your changes here and then delete this line
+### Added
+
+* Added `MemoryCacheHandler`, a cache handler that simply stores the token info in memory as an instance attribute of this class.
 
 ## [2.18.0] - 2021-04-13
